+++ process/coordinator/process.go
@@ -324,7 +324,7 @@ func (tc *transactionCoordinator) RemoveBlockDataFromPool(body block.Body) error
 }
 
 // ProcessBlockTransaction processes transactions and updates state tries
-func (tc *transactionCoordinator) ProcessBlockTransaction(body block.Body, round uint32, haveTime func() time.Duration) error {
+func (tc *transactionCoordinator) ProcessBlockTransaction(body block.Body, round uint64, haveTime func() time.Duration) error {
 	separatedBodies := tc.separateBodyByType(body)
 
 	var errFound error
@@ -365,7 +365,7 @@ func (tc *transactionCoordinator) CreateMbsAndProcessCrossShardTransactionsDstMe
 	hdr data.HeaderHandler,
 	maxTxRemaining uint32,
 	maxMbRemaining uint32,
-	round uint32,
+	round uint64,
 	haveTime func() bool,
 ) (block.MiniBlockSlice, uint32, bool) {
 	miniBlocks := make(block.MiniBlockSlice, 0)
@@ -438,7 +438,7 @@ func (tc *transactionCoordinator) CreateMbsAndProcessCrossShardTransactionsDstMe
 func (tc *transactionCoordinator) CreateMbsAndProcessTransactionsFromMe(
 	maxTxSpaceRemained uint32,
 	maxMbSpaceRemained uint32,
-	round uint32,
+	round uint64,
 	haveTime func() bool,
 ) block.MiniBlockSlice {
 
@@ -636,7 +636,7 @@ func (tc *transactionCoordinator) receivedMiniBlock(miniBlockHash []byte) {
 func (tc *transactionCoordinator) processCompleteMiniBlock(
 	preproc process.PreProcessor,
 	miniBlock *block.MiniBlock,
-	round uint32,
+	round uint64,
 	haveTime func() bool,
 ) error {
